requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.26.5